+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2025-10-19
+
+### Added
+
+- Added authentication support (Basic, Bearer, API Key, Digest)
+
+### Fixed
+
+- Fixed issue where whitespace was allowed in the **options-timeout** input field.
+
 ## [0.2.1] - 2025-10-18
 
 ### Fixed